cpu: rv64: conv: wino: move kernels from resource_t
